==> Sina Webo Interaction Prediction/9. Models.ipynb <==
import pandas as pd
import numpy as np
from sklearn import linear_model
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from matplotlib import pyplot as plt
import statsmodels.api as sm

import import_ipynb
from evaluation import precision
from runTime import runTime

from evaluation2 import precision2

df1=pd.read_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_train1_cp.csv")
df2=pd.read_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_train2_cp.csv")
frames=[df1,df2]
train_dataset=pd.concat(frames)
predict_dataset=pd.read_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_predict_cp.csv")

train_dataset['date']=pd.to_datetime(train_dataset['date'],errors='coerce')

train_month=[g for n, g in train_dataset.groupby(pd.Grouper(key='date',freq='M'))]

train_dataset['time']=pd.to_datetime(train_dataset['time'],errors='coerce')

train_hour=[g for n, g in train_dataset.groupby(pd.Grouper(key='time',freq='H'))]

train_month[0].to_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_train_feb_cp.csv",sep=',',index=False,encoding='utf-8')
train_month[1].to_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_train_march_cp.csv",sep=',',index=False,encoding='utf-8')
train_month[2].to_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_train_april_cp.csv",sep=',',index=False,encoding='utf-8')
train_month[3].to_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_train_may_cp.csv",sep=',',index=False,encoding='utf-8')
train_month[4].to_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_train_june_cp.csv",sep=',',index=False,encoding='utf-8')
train_month[5].to_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_train_july_cp.csv",sep=',',index=False,encoding='utf-8')

i=0
for i in range(0,24):
    path="E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_train_hour_"+str(i)+"_cp.csv"
    train_hour[i].to_csv(path,sep=',',index=False,encoding='utf-8')


frames1=[train_month[0],train_month[1],train_month[2],train_month[3],train_month[4]]
train=pd.concat(frames1)
predict=train_month[5]

# Model 1 (Factors: Media, #, @, ?, !, Length, Emoji)

X_train=train[["content_media_count","content_#_count","content_@_count","content_?_count","content_!_count","content_length","content_emoji_count"]]
Y_train=train[["forward_count","comment_count","like_count"]]
X_test=predict[["content_media_count","content_#_count","content_@_count","content_?_count","content_!_count","content_length","content_emoji_count"]]
Y_test=predict[["forward_count","comment_count","like_count"]]

print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred=lm.predict(X_test)
pred=pred.round()
pred=(np.maximum(pred,0.))

print(pred[0:5])

print(model.coef_)

print(model.intercept_)

np.savetxt("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_predict_result2.csv",pred,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result=pd.read_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_predict_result2.csv")

train_real_pred = Y_test
forward=result['forward_count'].values
comment=result['forward_count'].values
like=result['forward_count'].values
train_real_pred['fp'],train_real_pred['cp'],train_real_pred['lp'] = forward,comment,like
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model 2 (Media, Length, Emoji)

X_train=train[["content_media_count","content_length","content_emoji_count"]]
Y_train=train[["forward_count","comment_count","like_count"]]
X_test=predict[["content_media_count","content_length","content_emoji_count"]]
Y_test=predict[["forward_count","comment_count","like_count"]]

print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred=lm.predict(X_test)
pred=pred.round()
pred=(np.maximum(pred,0.))

print(pred[0:5])

print(model.coef_)

print(model.intercept_)

np.savetxt("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_predict_result3.csv",pred,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result=pd.read_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_predict_result3.csv")

train_real_pred = Y_test
forward=result['forward_count'].values
comment=result['forward_count'].values
like=result['forward_count'].values
train_real_pred['fp'],train_real_pred['cp'],train_real_pred['lp'] = forward,comment,like
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model 3(Media)

X_train=train[["content_media_count"]]
Y_train=train[["forward_count","comment_count","like_count"]]
X_test=predict[["content_media_count"]]
Y_test=predict[["forward_count","comment_count","like_count"]]

print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred=lm.predict(X_test)
pred=pred.round()
pred=(np.maximum(pred,0.))

print(pred[0:5])

print(model.coef_)

print(model.intercept_)

np.savetxt("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_predict_result4.csv",pred,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result=pd.read_csv("E:\\5th Sem\\DMA Project\\Model Evaluation\\weibo_predict_result4.csv")

train_real_pred = Y_test
train_real_pred['fp']=result['forward_count'].values
train_real_pred['cp']=result['comment_count'].values
train_real_pred['lp']=result['like_count'].values
train_real_pred=train_real_pred.round()
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model 4 (Time) Pre-requisite

train_dataset['hour']=pd.DatetimeIndex(train_dataset['time']).hour
print("done")
train_dataset['min']=pd.DatetimeIndex(train_dataset['time']).minute
print("done")
train_dataset['sec']=pd.DatetimeIndex(train_dataset['time']).second
print("done")

predict_dataset['hour']=pd.DatetimeIndex(predict_dataset['time']).hour
print("done")
predict_dataset['min']=pd.DatetimeIndex(predict_dataset['time']).minute
print("done")
predict_dataset['sec']=pd.DatetimeIndex(predict_dataset['time']).second
print("done")

train_dataset.head(614809).to_csv("G://DMA_PROJECT//weibo_train1_cpt.csv",sep=',',index=False,encoding='utf-8')
train_dataset.tail(614809).to_csv("G://DMA_PROJECT//weibo_train2_cpt.csv",sep=',',index=False,encoding='utf-8')
predict_dataset.to_csv("G://DMA_PROJECT//weibo_predict_cpt.csv",sep=',',index=False,encoding='utf-8')

df1=pd.read_csv("G://DMA_PROJECT//weibo_train1_cpt.csv")
df2=pd.read_csv("G://DMA_PROJECT//weibo_train2_cpt.csv")
frames=[df1,df2]
train_dataset=pd.concat(frames)
predict_dataset=pd.read_csv("G://DMA_PROJECT//weibo_predict_cpt.csv")

train_dataset['date']=pd.to_datetime(train_dataset['date'],errors='coerce')

train_month=[g for n, g in train_dataset.groupby(pd.Grouper(key='date',freq='M'))]

train_month[0].to_csv("G://DMA_PROJECT//weibo_train_feb_cpt.csv",sep=',',index=False,encoding='utf-8')
train_month[1].to_csv("G://DMA_PROJECT//weibo_train_march_cpt.csv",sep=',',index=False,encoding='utf-8')
train_month[2].to_csv("G://DMA_PROJECT//weibo_train_april_cpt.csv",sep=',',index=False,encoding='utf-8')
train_month[3].to_csv("G://DMA_PROJECT//weibo_train_may_cpt.csv",sep=',',index=False,encoding='utf-8')
train_month[4].to_csv("G://DMA_PROJECT//weibo_train_june_cpt.csv",sep=',',index=False,encoding='utf-8')
train_month[5].to_csv("G://DMA_PROJECT//weibo_train_july_cpt.csv",sep=',',index=False,encoding='utf-8')

train_month[0]=pd.read_csv("G://DMA_PROJECT//weibo_train_feb_cpt.csv")
train_month[1]=pd.read_csv("G://DMA_PROJECT//weibo_train_march_cpt.csv")
train_month[2]=pd.read_csv("G://DMA_PROJECT//weibo_train_april_cpt.csv")
train_month[3]=pd.read_csv("G://DMA_PROJECT//weibo_train_may_cpt.csv")
train_month[4]=pd.read_csv("G://DMA_PROJECT//weibo_train_june_cpt.csv")
train_month[5]=pd.read_csv("G://DMA_PROJECT//weibo_train_july_cpt.csv")

frames1=[train_month[0],train_month[1],train_month[2],train_month[3],train_month[4]]
train=pd.concat(frames1)
predict=train_month[5]

# Model 4 (Time: Hour, Min, Sec)

X_train=train[["hour","min","sec"]]
Y_train=train[["forward_count","comment_count","like_count"]]
X_test=predict[["hour","min","sec"]]
Y_test=predict[["forward_count","comment_count","like_count"]]

print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred=lm.predict(X_test)
pred=pred.round()
pred=(np.maximum(pred,0.))

print(pred[0:5])

print(model.coef_)

print(model.intercept_)

np.savetxt("G://DMA_PROJECT//weibo_predict_result5.csv",pred,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result=pd.read_csv("G://DMA_PROJECT//weibo_predict_result5.csv")

train_real_pred = Y_test
train_real_pred['fp']=result['forward_count'].values
train_real_pred['cp']=result['comment_count'].values
train_real_pred['lp']=result['like_count'].values
train_real_pred=train_real_pred.round()
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model 5 (Time: Hour)

X_train=train[["hour"]]
Y_train=train[["forward_count","comment_count","like_count"]]
X_test=predict[["hour"]]
Y_test=predict[["forward_count","comment_count","like_count"]]

print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred=lm.predict(X_test)
pred=pred.round()
pred=(np.maximum(pred,0.))

print(pred[0:5])

print(model.coef_)

print(model.intercept_)

np.savetxt("G://DMA_PROJECT//weibo_predict_result6.csv",pred,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result=pd.read_csv("G://DMA_PROJECT//weibo_predict_result6.csv")

train_real_pred = Y_test
train_real_pred['fp']=result['forward_count'].values
train_real_pred['cp']=result['comment_count'].values
train_real_pred['lp']=result['like_count'].values
train_real_pred=train_real_pred.round()
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model 6  Time: (Hour, Min, Sec), Media, Length, Emoji

X_train=train[["content_media_count","content_length","content_emoji_count","hour","min","sec"]]
Y_train=train[["forward_count","comment_count","like_count"]]
X_test=predict[["content_media_count","content_length","content_emoji_count","hour","min","sec"]]
Y_test=predict[["forward_count","comment_count","like_count"]]

print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred=lm.predict(X_test)
pred=pred.round()
pred=(np.maximum(pred,0.))

print(pred[0:5])

print(model.coef_)

print(model.intercept_)

np.savetxt("G://DMA_PROJECT//weibo_predict_result7.csv",pred,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result=pd.read_csv("G://DMA_PROJECT//weibo_predict_result7.csv")

train_real_pred = Y_test
train_real_pred['fp']=result['forward_count'].values
train_real_pred['cp']=result['comment_count'].values
train_real_pred['lp']=result['like_count'].values
train_real_pred=train_real_pred.round()
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model 7: Stats Pre-requisite

df1=pd.read_csv("G://DMA_PROJECT//weibo_train1_cpt.csv")
df2=pd.read_csv("G://DMA_PROJECT//weibo_train2_cpt.csv")
frames=[df1,df2]
train_dataset=pd.concat(frames)
predict_dataset=pd.read_csv("G://DMA_PROJECT//weibo_predict_cpt.csv")

stat=pd.read_csv("G://DMA_PROJECT//train_uid_stat.csv")

trainstat=pd.merge(train_dataset,stat,on=['u_id'])
predictstat=pd.merge(train_dataset,stat,on=['u_id'])

trainstat.head(5)

trainstat.head(614809).to_csv("G://DMA_PROJECT//weibo_train1_cpts.csv",sep=',',index=False,encoding='utf-8')
trainstat.tail(614809).to_csv("G://DMA_PROJECT//weibo_train2_cpts.csv",sep=',',index=False,encoding='utf-8')
predictstat.to_csv("G://DMA_PROJECT//weibo_predict_cpts.csv",sep=',',index=False,encoding='utf-8')

df1=pd.read_csv("G://DMA_PROJECT//weibo_train1_cpts.csv")
df2=pd.read_csv("G://DMA_PROJECT//weibo_train2_cpts.csv")
frames=[df1,df2]
train_dataset=pd.concat(frames)
predict_dataset=pd.read_csv("G://DMA_PROJECT//weibo_predict_cpts.csv")

train_dataset['date']=pd.to_datetime(train_dataset['date'],errors='coerce')

train_month=[g for n, g in train_dataset.groupby(pd.Grouper(key='date',freq='M'))]

train_month[0].to_csv("G://DMA_PROJECT//weibo_train_feb_cpts.csv",sep=',',index=False,encoding='utf-8')
train_month[1].to_csv("G://DMA_PROJECT//weibo_train_march_cpts.csv",sep=',',index=False,encoding='utf-8')
train_month[2].to_csv("G://DMA_PROJECT//weibo_train_april_cpts.csv",sep=',',index=False,encoding='utf-8')
train_month[3].to_csv("G://DMA_PROJECT//weibo_train_may_cpts.csv",sep=',',index=False,encoding='utf-8')
train_month[4].to_csv("G://DMA_PROJECT//weibo_train_june_cpts.csv",sep=',',index=False,encoding='utf-8')
train_month[5].to_csv("G://DMA_PROJECT//weibo_train_july_cpts.csv",sep=',',index=False,encoding='utf-8')

train_month[0]=pd.read_csv("G://DMA_PROJECT//weibo_train_feb_cpts.csv")
train_month[1]=pd.read_csv("G://DMA_PROJECT//weibo_train_march_cpts.csv")
train_month[2]=pd.read_csv("G://DMA_PROJECT//weibo_train_april_cpts.csv")
train_month[3]=pd.read_csv("G://DMA_PROJECT//weibo_train_may_cpts.csv")
train_month[4]=pd.read_csv("G://DMA_PROJECT//weibo_train_june_cpts.csv")
train_month[5]=pd.read_csv("G://DMA_PROJECT//weibo_train_july_cpts.csv")

frames1=[train_month[0],train_month[1],train_month[2],train_month[3],train_month[4]]
train=pd.concat(frames1)
predict=train_month[5]

# Model 7 Median,Time: (Hour, Min, Sec), Media, Length, Emoji 
## Only for Forward Count

X_train=train[["content_media_count","content_length","content_emoji_count","hour","min","sec","forward_median"]]
Y_train=train[["forward_count"]]
X_test=predict[["content_media_count","content_length","content_emoji_count","hour","min","sec","forward_median"]]
Y_test=predict[["forward_count"]]

print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred=lm.predict(X_test)
pred=pred.round()
pred=(np.maximum(pred,0.))

print(pred[0:5])

print(model.coef_)

print(model.intercept_)

np.savetxt("G://DMA_PROJECT//weibo_predict_result8.csv",pred,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result=pd.read_csv("G://DMA_PROJECT//weibo_predict_result8.csv")

train_real_pred=Y_test
train_real_pred['fp']=result['forward_count'].values
train_real_pred=train_real_pred.round()
print ("Score on the training set:{0:.2f}%".format(precision2(train_real_pred.values)*100))

# Model 7 Median,Time: (Hour, Min, Sec), Media, Length, Emoji 

X_train=train[["content_media_count","content_length","content_emoji_count","hour","min","sec","forward_median","comment_median","like_median"]]
Y_train=train[["forward_count","comment_count","like_count"]]
X_test=predict[["content_media_count","content_length","content_emoji_count","hour","min","sec","forward_median","comment_median","like_median"]]
Y_test=predict[["forward_count","comment_count","like_count"]]

print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred=lm.predict(X_test)
pred=pred.round()
pred=(np.maximum(pred,0.))

print(pred[0:5])

print(model.coef_)

print(model.intercept_)

np.savetxt("G://DMA_PROJECT//weibo_predict_result9.csv",pred,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result=pd.read_csv("G://DMA_PROJECT//weibo_predict_result9.csv")

train_real_pred = Y_test
train_real_pred['fp']=result['forward_count'].values
train_real_pred['cp']=result['comment_count'].values
train_real_pred['lp']=result['like_count'].values
train_real_pred=train_real_pred.round()
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))
