==> Sina Webo Interaction Prediction/11. Ramdom Forest Regressor.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns 
%pylab inline
import copy
from googletrans import Translator
import pandas as pd
import numpy as np
import csv
import re
import jieba
import time
import json
from sklearn.feature_extraction.text import CountVectorizer
from sklearn import linear_model
from sklearn.externals import joblib
from nltk.corpus import stopwords as e_stopwords
from datetime import datetime, timedelta
import jieba
import sys

from nltk.corpus import stopwords 
from nltk.tokenize import word_tokenize

# Random Forest Regressor for Statistical Factors

import import_ipynb
from evaluation import precision
from runTime import runTime

#Reading data from document
import pandas as pd
df_pre=pd.read_csv("E:\\DMA_PRE\\PREPROCESSED.csv")

df_pre.shape[0]

train=df_pre[0:8000]
cv=df_pre[8001:10000]

df_stat=pd.read_csv("E:\\DMA_PRE\\train_uid_stat.csv")


df_stat.shape

df_stat.head(5)

train_all1=pd.read_csv('E:\\5th-Sem\\DMA Project\\Model Evaluation\\weibo_train1_cp.csv')
train_all2=pd.read_csv('E:\\5th-Sem\\DMA Project\\Model Evaluation\\weibo_train2_cp.csv')
frames=[train_all1,train_all2]
train_all=pd.concat(frames)

train_all.shape[0]

df_merge = pd.merge(train_all, df_stat, how='left', on=['u_id'])

df1=pd.read_csv("E:\\5th-Sem\\DMA Project\\Project\\weibo_train1_cpts.csv")
df2=pd.read_csv("E:\\5th-Sem\\DMA Project\\Project\\weibo_train2_cpts.csv")
frames=[df1,df2]
train_all=pd.concat(frames)

X=train_all[["content_media_count","content_#_count","content_length","content_emoji_count","forward_median","comment_median","like_median"]]
y=train_all[['forward_count', 'comment_count', 'like_count']]

from sklearn import cross_validation
## Spliting of training dataset into 70% training data and 30% testing data randomly
features_train, features_test, labels_train, labels_test = cross_validation.train_test_split(X, y, test_size=0.3, random_state=1)



# Model 1 (Predicting all 3 values together)

from sklearn.ensemble import RandomForestRegressor

x = features_train
y = labels_train
x1 = features_test
y1 = labels_test



regr = RandomForestRegressor(max_depth=30, random_state=0,n_estimators=100)
regr.fit(x, y)
y11_predict = regr.predict(x1)
y11_predict=y11_predict.round()
y11_predict=(np.maximum(y11_predict,0.))
#print(r2_score(y1, y11_predict) ) #Random forest regressor

np.savetxt("E:\\weibo_predict_result.csv",y11_predict,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result=pd.read_csv("E:\\weibo_predict_result.csv")


train_real_pred = labels_test
forward=result['forward_count'].values
comment=result['comment_count'].values
like=result['like_count'].values
train_real_pred['fp'],train_real_pred['cp'],train_real_pred['lp'] = forward,comment,like
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))

train_all.head(1).T

# Model 2 - Constructing 3 individual models and concatinating the results

X=train_all[["forward_median","forward_mean","forward_min","content_media_count","content_emoji_count"]]
y=train_all['forward_count']
from sklearn import cross_validation
## Spliting of training dataset into 70% training data and 30% testing data randomly
features_train, features_test, labels_train, labels_test = cross_validation.train_test_split(X, y, test_size=0.3, random_state=1)
from sklearn.ensemble import RandomForestRegressor

x = features_train
y = labels_train
x1 = features_test
y1 = labels_test



regr = RandomForestRegressor(max_depth=30, random_state=0,n_estimators=100)
regr.fit(x, y)
y11_predict = regr.predict(x1)
#print(r2_score(y1, y11_predict) ) #Random forest regressor
np.savetxt("E:\\weibo_predict_result1.csv",y11_predict,delimiter=',',header="forward_count",comments="")
result1=pd.read_csv("E:\\weibo_predict_result1.csv")


X=train_all[["comment_median","comment_mean","comment_min","content_media_count","content_emoji_count"]]
y=train_all['comment_count']
from sklearn import cross_validation
## Spliting of training dataset into 70% training data and 30% testing data randomly
features_train, features_test, labels_train, labels_test = cross_validation.train_test_split(X, y, test_size=0.3, random_state=1)
from sklearn.ensemble import RandomForestRegressor

x = features_train
y = labels_train
x1 = features_test
y2 = labels_test



regr = RandomForestRegressor(max_depth=30, random_state=0,n_estimators=100)
regr.fit(x, y)
y11_predict = regr.predict(x1)
#print(r2_score(y1, y11_predict) ) #Random forest regressor
np.savetxt("E:\\weibo_predict_result2.csv",y11_predict,delimiter=',',header="comment_count",comments="")
result2=pd.read_csv("E:\\weibo_predict_result2.csv")

X=train_all[["like_median","like_mean","like_min","content_media_count","content_emoji_count"]]
y=train_all['like_count']
from sklearn import cross_validation
## Spliting of training dataset into 70% training data and 30% testing data randomly
features_train, features_test, labels_train, labels_test = cross_validation.train_test_split(X, y, test_size=0.3, random_state=1)
from sklearn.ensemble import RandomForestRegressor

x = features_train
y = labels_train
x1 = features_test
y3 = labels_test



regr = RandomForestRegressor(max_depth=30, random_state=0,n_estimators=100)
regr.fit(x, y)
y11_predict = regr.predict(x1)
#print(r2_score(y1, y11_predict) ) #Random forest regressor
np.savetxt("E:\\weibo_predict_result3.csv",y11_predict,delimiter=',',header="like_count",comments="")
result3=pd.read_csv("E:\\weibo_predict_result3.csv")

train_real_pred = pd.concat([y1,y2,y3],axis=1)
train_real_pred['fp']=result1['forward_count']
train_real_pred['cp']=result2['comment_count']
train_real_pred['lp']=result3['like_count']
train_real_pred=train_real_pred.round()
print ("Score on the training set:{0:.4f}%".format(precision(train_real_pred.values)*100))
