==> Sina Webo Interaction Prediction/12. Generate Best Prediction.ipynb <==
import pandas as pd
import csv

import _pickle as cPickle
import import_ipynb
import pandas as pd
import numpy as np
from genUidStat import loadData,genUidStat
from evaluation import precision
from runTime import runTime
from pathos.pools import _ProcessPool
from multiprocess.pool import Pool

df1=pd.read_csv("weibo_train1.csv")
df2=pd.read_csv("weibo_train2.csv")
frames=[df1,df2]
traindata=pd.concat(frames)

def splitDataFrameIntoSmaller(df, chunkSize = 10000): 
    listOfDf = list()
    numberChunks = len(df) // chunkSize + 1
    for i in range(numberChunks):
        listOfDf.append(df[i*chunkSize:(i+1)*chunkSize])
    return listOfDf

uid_stat=pd.read_csv("train_uid_stat.csv")

uid = splitDataFrameIntoSmaller(uid_stat, chunkSize = 500)

uid[0].shape[0]

def search_all_uid(stat_dic,file):
	import pandas as pd
	import numpy as np
	def _deviation(predict, real, kind):
		t = 5.0 if kind=='f' else 3.0
		return abs(predict - real) / (real + t)
	def _precision_i(fp, fr, cp, cr, lp, lr):
		return 1 - 0.5 * _deviation(fp, fr, 'f') - 0.25 * _deviation(cp, cr, 'c') - 0.25 * _deviation(lp, lr, 'l')
	def _sgn(x):
		return 1 if x>0 else 0
	def _count_i(fr, cr, lr):
		x = fr + cr + lr
		return 101 if x>100 else (x+1)
	def precision(real_and_predict):
		numerator,denominator = 0.0,0.0
		for  fr, cr, lr,fp, cp, lp in real_and_predict:
			numerator += _count_i(fr, cr, lr) * _sgn(_precision_i(fp, fr, cp, cr, lp, lr) - 0.8)
			denominator += _count_i(fr, cr, lr)
		return (numerator / denominator)
	def score(uid_data,pred):
		"""
		uid_data:
			pd.DataFrame
		pred:
			list, [fp,cp,lp]
		"""
		uid_real_pred = uid_data[['forward_count','comment_count','like_count']]
		uid_real_pred['fp'] = pred[0]
		uid_real_pred['cp'] = pred[1]
		uid_real_pred['lp'] = pred[2]
		return precision(uid_real_pred.values)
	def search(uid_data,target,args):
		args = list(args)
		target_index = ['forward_count','comment_count','like_count'].index(target)
		target_min,target_median,target_max = args[3*target_index:3*target_index+3]
		del args[3*target_index:3*target_index+3]
		pred = (args[1],args[4])
		
		best_num = [target_median]
		best_pred = list(pred)
		best_pred.insert(target_index,target_median)
		best_score = score(uid_data,best_pred)
		for num in range(target_min,target_max+1):
			this_pred = list(pred)
			this_pred.insert(target_index,num)
			this_score = score(uid_data,this_pred)
			if this_score >= best_score:                  
				if this_score > best_score:
					best_num = [num]
					best_score = this_score
				else:
					best_num.append(num)                       
				
		return best_num[np.array([abs(i - target_median) for i in best_num]).argmin()]
	uid_best_pred = {}
	pool = _ProcessPool()
	uids,f,c,l = [],[],[],[]
	m=1
	for uid in stat_dic:
		print ("search uid:{}".format(uid),m)
		m=m+1
		uid_data = traindata[traindata.u_id == uid]
		arguments = stat_dic[uid][['forward_min','forward_median','forward_max','comment_min',\
					'comment_median','comment_max','like_min','like_median','like_max']]
		arguments = tuple([int(i) for i in arguments]) 
		f.append(pool.apply_async(search,args=(uid_data,'forward_count',arguments)))
		c.append(pool.apply_async(search,args=(uid_data,'comment_count',arguments)))
		l.append(pool.apply_async(search,args=(uid_data,'like_count',arguments)))
		uids.append(uid)
	pool.close()
	pool.join()
	f = [i.get() for i in f]
	c = [i.get() for i in c]
	l = [i.get() for i in l]
	for i in range(len(uids)):
		uid_best_pred[uids[i]] = [f[i],c[i],l[i]]
	#cPickle.dump(uid_best_pred,open('uid_best_pred'+str(file)+'.pkl','ab'))
	label = ['forward_count','comment_count','like_count']
	pd.DataFrame.from_dict(data=uid_best_pred,orient='index').to_csv("G:\\Anconda Prog\\BestPred\\weibo_uidbest"+str(file)+".csv",header=label)
	print("Written to file")



uid_stat=pd.read_csv("train_uid_stat.csv")
uid_stat=uid_stat.set_index('u_id')
uid = splitDataFrameIntoSmaller(uid_stat, chunkSize = 100)
n=8
while n<75:
    df=uid[n].T
    stat=df.to_dict('series')
    n=n+1
    search_all_uid(stat,n)

# FILES GENERATED

@runTime
def predict_by_search(submission=True):
	traindata,testdata = loadData()
	#concat all frames here
	
	
	ub=pd.read_csv("train_best_pred.csv")
	ub=ub.set_index('u_id')
	df=ub.T
	uid_best_pred=df.to_dict('series')
	#uid_best_pred = search_all_uid()
	#print ("search done,now predict on traindata and testdata...")

	#predict traindata with uid's best fp,cp,lp
	forward,comment,like = [],[],[]
	for uid in traindata['u_id']:
		if uid in uid_best_pred:
			forward.append(int(uid_best_pred[uid][0]))
			comment.append(int(uid_best_pred[uid][1]))
			like.append(int(uid_best_pred[uid][2]))
		else:
			forward.append(0)
			comment.append(0)
			like.append(0)
	#score on the traindata
	train_real_pred = traindata[['forward_count','comment_count','like_count']]
	train_real_pred['fp'],train_real_pred['cp'],train_real_pred['lp'] = forward,comment,like
	print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))
	if submission:
		test_pred = testdata[['u_id','m_id']]
		forward,comment,like = [],[],[]
		for uid in testdata['u_id']:
			if uid in uid_best_pred:
				forward.append(int(uid_best_pred[uid][0]))
				comment.append(int(uid_best_pred[uid][1]))
				like.append(int(uid_best_pred[uid][2]))
			else:
				forward.append(0)
				comment.append(0)
				like.append(0)
		test_pred['fp'],test_pred['cp'],test_pred['lp'] = forward,comment,like
		
		#generate submission file
		result = []
		filename = "weibo_predict_search.txt"
		for _,row in test_pred.iterrows():
			result.append("{0}\t{1}\t{2},{3},{4}\n".format(row[0],row[1],row[2],row[3],row[4]))
		f = open(filename,'w')
		f.writelines(result)
		f.close()
		print ('generate submission file "{}"'.format(filename))

predict_by_search()
