==> Sina Webo Interaction Prediction/13. Model Evaluation(OLS).ipynb <==
import pandas as pd
import numpy as np
from sklearn import linear_model
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from matplotlib import pyplot as plt
import statsmodels.api as sm
import import_ipynb
from evaluation import precision

train_month1=pd.read_csv("E:\5th Sem\DMA Project\Model Evaluation\weibo_train_feb_cpts.csv")
train_month2=pd.read_csv("E:\5th Sem\DMA Project\Model Evaluation\weibo_train_march_cpts.csv")
train_month3=pd.read_csv("E:\5th Sem\DMA Project\Model Evaluation\weibo_train_april_cpts.csv")
train_month4=pd.read_csv("E:\5th Sem\DMA Project\Model Evaluation\weibo_train_may_cpts.csv")
train_month5=pd.read_csv("E:\5th Sem\DMA Project\Model Evaluation\weibo_train_june_cpts.csv")
train_month6=pd.read_csv("E:\5th Sem\DMA Project\Model Evaluation\weibo_train_july_cpts.csv")

frames1=[train_month1,train_month2,train_month3,train_month4,train_month5]
train=pd.concat(frames1)
predict=train_month6

X_train1=train[["forward_median","forward_mean","forward_min","content_media_count","content_emoji_count"]]
Y_train1=train[["forward_count"]]
X_test1=predict[["forward_median","forward_mean","forward_min","content_media_count","content_emoji_count"]]
Y_test1=predict[["forward_count"]]

X_train2=train[["comment_median","comment_mean","comment_min","content_media_count","content_emoji_count"]]
Y_train2=train[["comment_count"]]
X_test2=predict[["comment_median","comment_mean","comment_min","content_media_count","content_emoji_count"]]
Y_test2=predict[["comment_count"]]

X_train3=train[["like_median","like_mean","like_min","content_media_count","content_emoji_count"]]
Y_train3=train[["like_count"]]
X_test3=predict[["like_median","like_mean","like_min","content_media_count","content_emoji_count"]]
Y_test3=predict[["like_count"]]

pd.options.mode.use_inf_as_na = True
X_train1.fillna(X_train1.max(),inplace=True)
X_test1.fillna(X_test1.max(),inplace=True)
X_train2.fillna(X_train2.max(),inplace=True)
X_test2.fillna(X_test2.max(),inplace=True)
X_train3.fillna(X_train3.max(),inplace=True)
X_test3.fillna(X_test3.max(),inplace=True)

print(X_train1.shape,Y_train1.shape)
print(X_test1.shape,Y_test1.shape)

print(X_train2.shape,Y_train2.shape)
print(X_test2.shape,Y_test2.shape)

print(X_train3.shape,Y_train3.shape)
print(X_test3.shape,Y_test3.shape)

model1=sm.OLS(Y_train1,X_train1).fit()
pred1=model1.predict(X_test1)
pred1=pred1.round()
pred1=(np.maximum(pred1,0.))
print(model1.summary())

model2=sm.OLS(Y_train2,X_train2).fit()
pred2=model2.predict(X_test2)
pred2=pred2.round()
pred2=(np.maximum(pred2,0.))
print(model2.summary())

model3=sm.OLS(Y_train3,X_train3).fit()
pred3=model3.predict(X_test3)
pred3=pred3.round()
pred3=(np.maximum(pred3,0.))
print(model3.summary())

print(pred1[0:5])

print(pred2[0:5])

print(pred3[0:5])

np.savetxt("G://DMA_PROJECT//weibo_predict_resulto1.csv",pred1,delimiter=',',header="forward_count",comments="")
result1=pd.read_csv("G://DMA_PROJECT//weibo_predict_resulto1.csv")
np.savetxt("G://DMA_PROJECT//weibo_predict_resulto2.csv",pred2,delimiter=',',header="comment_count",comments="")
result2=pd.read_csv("G://DMA_PROJECT//weibo_predict_resulto2.csv")
np.savetxt("G://DMA_PROJECT//weibo_predict_resulto3.csv",pred3,delimiter=',',header="like_count",comments="")
result3=pd.read_csv("G://DMA_PROJECT//weibo_predict_resulto3.csv")

train_real_pred = pd.concat([Y_test1,Y_test2,Y_test3],axis=1)
train_real_pred['fp']=result1['forward_count']
train_real_pred['cp']=result2['comment_count']
train_real_pred['lp']=result3['like_count']
train_real_pred=train_real_pred.round()
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))
