==> Sina Webo Interaction Prediction/14. Ensemble - Averaging.ipynb <==
import pandas as pd
import numpy as np
import re
from sklearn import linear_model
from sklearn.linear_model import Lasso
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error
from matplotlib import pyplot as plt
from textblob import TextBlob
import statsmodels.api as sm

import import_ipynb
from evaluation import precision
from runTime import runTime

df1=pd.read_csv("E:\\5th-Sem\\DMA Project\\Project\\weibo_train1_cpts.csv")
df2=pd.read_csv("E:\\5th-Sem\\DMA Project\\Project\\weibo_train2_cpts.csv")
frames=[df1,df2]
train_all=pd.concat(frames)

train_all.shape[0]

train=train_all[0:983694]
predict=train_all[983695:1229618]

train.columns

# Model 1 - Linear Regression


X_train=train[["content_media_count","content_length","content_emoji_count","hour","min","sec","forward_median","comment_median","like_median"]]
Y_train=train[["forward_count","comment_count","like_count"]]
X_test=predict[["content_media_count","content_length","content_emoji_count","hour","min","sec","forward_median","comment_median","like_median"]]
Y_test=predict[["forward_count","comment_count","like_count"]]
print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred1=lm.predict(X_test)


temp = pd.DataFrame.from_records(pred1)
temp=temp.round()
temp=(np.maximum(temp,0))
train_real_pred=Y_test
train_real_pred['fp']=temp[0].values
train_real_pred['cp']=temp[1].values
train_real_pred['lp']=temp[2].values
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))

from sklearn.model_selection import KFold
kf = KFold(n_splits=10)
kf.get_n_splits(X)  
KFold(n_splits=10, random_state=None, shuffle=False)
for train_index, test_index in kf.split(X):
print("TRAIN:", train_index, "TEST:", test_index)
X_train, X_test = X[train_index], X[test_index]
y_train, y_test = y[train_index], y[test_index]

# Model 2 - Random Forest

from sklearn.ensemble import RandomForestRegressor

## Spliting of training dataset into 70% training data and 30% testing data randomly
features_train=train[["content_media_count","content_#_count","content_length","content_emoji_count","forward_median","comment_median","like_median"]]
features_test=predict[["content_media_count","content_#_count","content_length","content_emoji_count","forward_median","comment_median","like_median"]]
labels_train=train[['forward_count', 'comment_count', 'like_count']]
labels_test=predict[['forward_count', 'comment_count', 'like_count']]

x = features_train
y = labels_train
x1 = features_test
y1 = labels_test



regr = RandomForestRegressor(max_depth=50, random_state=0,n_estimators=100)
regr.fit(x, y)
pred2 = regr.predict(x1)

temp = pd.DataFrame.from_records(pred2)
temp=temp.round()
temp=(np.maximum(temp,0))
temp=temp.abs()
temp=temp.astype(int)
train_real_pred=Y_test
train_real_pred['fp']=temp[0].values
train_real_pred['cp']=temp[1].values
train_real_pred['lp']=temp[2].values
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))

#train_real_pred

# Model 3 - OLS

model3=sm.OLS(Y_train,X_train).fit()
pred3=model3.predict(X_test)

# Model 4- Ridge

lm1=linear_model.Ridge(alpha=3)
model4=lm1.fit(X_train,Y_train)
pred4=lm1.predict(X_test)

# Model 5- Lasso

lm1=Lasso(alpha=0.01)
model5=lm1.fit(X_train,Y_train)
pred5=lm1.predict(X_test)

# Ensemble - Averaging

pred=(pred1+pred2+pred3+pred4+pred5)/5

pred=pred.round()
pred=(np.maximum(pred,0))

pred=pred.abs()
pred1=pred.astype(int)

pred1

#np.savetxt("E://DMA_PRED//result_Ensemble.csv",pred1,delimiter=',',header="forward_count,comment_count,like_count",comments="")
#result1=pd.read_csv("E://DMA_PRED//result_Ensemble.csv")

#result1

train_real_pred=Y_test
train_real_pred['fp']=pred1[0]
train_real_pred['cp']=pred1[1]
train_real_pred['lp']=pred1[2]
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))

#train_real_pred

# XG Boost

x = features_train
y = labels_train
x1 = features_test
y1 = labels_test
import xgboost as xgb

#T_train_xgb = xgb.DMatrix(x, y)
#params = {"objective": "reg:linear", "booster":"gblinear"}
#gbm = xgb.train(dtrain=T_train_xgb,params=params)
#Y_pred = gbm.predict(xgb.DMatrix(x1))
#print(r2_score(y1, Y_pred) )  #xgboost

# Mapping Uid


unique_id=train_all['u_id'].unique().tolist()

uid_df = pd.DataFrame({'u_id':unique_id})

uid_df.shape[0]

from sklearn import preprocessing
le = preprocessing.LabelEncoder()
le.fit(unique_id)
#list(le.classes_)


l=[]
l=le.transform(unique_id) 
df = pd.DataFrame({'u_id':l})
uid_df['id']=df['u_id']

train_all=train_all.set_index('u_id').join(uid_df.set_index('u_id'))

train=train_all[0:983694]
predict=train_all[983695:1229618]

X_train=train[["id","content_media_count","content_length","content_emoji_count","hour","min","sec","forward_median","comment_median","like_median"]]
Y_train=train[["forward_count","comment_count","like_count"]]
X_test=predict[["id","content_media_count","content_length","content_emoji_count","hour","min","sec","fo
                rward_median","comment_median","like_median"]]
Y_test=predict[["forward_count","comment_count","like_count"]]
print(X_train.shape,Y_train.shape)
print(X_test.shape,Y_test.shape)

pd.options.mode.use_inf_as_na = True
X_train.fillna(X_train.max(),inplace=True)
X_test.fillna(X_test.max(),inplace=True)

lm=linear_model.LinearRegression()
model=lm.fit(X_train,Y_train)
pred1=lm.predict(X_test)


temp = pd.DataFrame.from_records(pred1)
temp=temp.round()
temp=(np.maximum(temp,0))
temp=temp.abs()
temp=temp.astype(int)
train_real_pred=Y_test
train_real_pred['fp']=temp[0].values
train_real_pred['cp']=temp[1].values
train_real_pred['lp']=temp[2].values
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))

## Spliting of training dataset into 70% training data and 30% testing data randomly
features_train=train[["id","content_media_count","content_#_count","content_length","content_emoji_count","forward_median","comment_median","like_median"]]
features_test=predict[["id","content_media_count","content_#_count","content_length","content_emoji_count","forward_median","comment_median","like_median"]]
labels_train=train[['forward_count', 'comment_count', 'like_count']]
labels_test=predict[['forward_count', 'comment_count', 'like_count']]

x = features_train
y = labels_train
x1 = features_test
y1 = labels_test



regr = RandomForestRegressor(max_depth=50, random_state=0,n_estimators=100)
regr.fit(x, y)
pred2 = regr.predict(x1)

temp = pd.DataFrame.from_records(pred2)
temp=temp.round()
temp=(np.maximum(temp,0))
temp=temp.abs()
temp=temp.astype(int)
train_real_pred=Y_test
train_real_pred['fp']=temp[0].values
train_real_pred['cp']=temp[1].values
train_real_pred['lp']=temp[2].values
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))
