==> Sina Webo Interaction Prediction/15. Polarity-Iteration 2.ipynb <==
import pandas as pd
import numpy as np
import re
from sklearn import linear_model
from sklearn.linear_model import Lasso
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error
from matplotlib import pyplot as plt
from textblob import TextBlob
import statsmodels.api as sm

import import_ipynb
from evaluation import precision
from runTime import runTime

# ------------------------------------------Polarity as a factor--------------------------------------

### Sentences with positive meaning have positive polarity and negative meaning ones have negative polarity and neutral ones have zero as polarity

dfpol=pd.read_csv("C:/Users/user/Downloads/weibo_polarity.csv")

dfpol.head(10)

dfpol.columns

dfpol['date']=pd.to_datetime(dfpol['date'],errors='coerce')
train_month=[g for n, g in dfpol.groupby(pd.Grouper(key='date',freq='M'))]

train_month[0]=pd.read_csv("C:/Users/user/Downloads/weibo_train_feb_cpts10000.csv")
train_month[1]=pd.read_csv("C:/Users/user/Downloads/weibo_train_march_cpts10000.csv")
train_month[2]=pd.read_csv("C:/Users/user/Downloads/weibo_train_april_cpts10000.csv")
train_month[3]=pd.read_csv("C:/Users/user/Downloads/weibo_train_may_cpts10000.csv")
train_month[4]=pd.read_csv("C:/Users/user/Downloads/weibo_train_june_cpts10000.csv")
train_month[5]=pd.read_csv("C:/Users/user/Downloads/weibo_train_july_cpts10000.csv")


frames1=[train_month[0],train_month[1],train_month[2],train_month[3],train_month[4]]
train=pd.concat(frames1)
predict=train_month[5]

## Model 7: (Factors: Media, Length, Emoji, Median,Polarity)

X_train1=train[["content_media_count","content_length","forward_median","comment_median","like_median","polarity"]]
Y_train1=train[["forward_count","comment_count","like_count"]]
X_test1=predict[["content_media_count","content_length","forward_median","comment_median","like_median","polarity"]]
Y_test1=predict[["forward_count","comment_count","like_count"]]

pd.options.mode.use_inf_as_na = True
X_train1.fillna(X_train1.max(),inplace=True)
X_test1.fillna(X_test1.max(),inplace=True)

lm1=linear_model.LinearRegression()
model1=lm1.fit(X_train1,Y_train1)
pred1=lm1.predict(X_test1)
pred1=pred1.round()
pred1=(np.maximum(pred1,0.))

print(model1.coef_)
print(model1.intercept_)

np.savetxt("C:/Users/user/Downloads/weibo_predict_result51.csv",pred1,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result1=pd.read_csv("C:/Users/user/Downloads/weibo_predict_result51.csv")

print(mean_squared_error(Y_test1,result1))

train_real_pred=Y_test1
train_real_pred['fp']=result1['forward_count']
train_real_pred['cp']=result1['comment_count']
train_real_pred['lp']=result1['like_count']
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))

## Analysis: Result with Polarity as factor are satisfactory considering the data used for train. This might prove to be a good factor for whole dataset prediction.

# Model 8: (Factors: Media, Length,Median,Polarity)

X_train1=train[["content_media_count","content_length","forward_mean","comment_mean","like_mean","polarity"]]
Y_train1=train[["forward_count","comment_count","like_count"]]
X_test1=predict[["content_media_count","content_length","forward_mean","comment_mean","like_mean","polarity"]]
Y_test1=predict[["forward_count","comment_count","like_count"]]

pd.options.mode.use_inf_as_na = True
X_train1.fillna(X_train1.max(),inplace=True)
X_test1.fillna(X_test1.max(),inplace=True)

lm1=linear_model.LinearRegression()
model1=lm1.fit(X_train1,Y_train1)
pred1=lm1.predict(X_test1)
pred1=pred1.round()
pred1=(np.maximum(pred1,0.))

print(model1.coef_)
print(model1.intercept_)

np.savetxt("C:/Users/user/Downloads/weibo_predict_result51.csv",pred1,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result1=pd.read_csv("C:/Users/user/Downloads/weibo_predict_result51.csv")

print(mean_squared_error(Y_test1,result1))

train_real_pred=Y_test1
train_real_pred['fp']=result1['forward_count']
train_real_pred['cp']=result1['comment_count']
train_real_pred['lp']=result1['like_count']
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model 9: (Factors: Media, Length, Emoji, Median,Mean,Polarity)

X_train1=train[["content_media_count","content_length","forward_mean","forward_median","comment_mean","comment_median","like_mean","like_median","polarity"]]
Y_train1=train[["forward_count","comment_count","like_count"]]
X_test1=predict[["content_media_count","content_length","forward_mean","forward_median","comment_mean","comment_median","like_mean","like_median","polarity"]]
Y_test1=predict[["forward_count","comment_count","like_count"]]

pd.options.mode.use_inf_as_na = True
X_train1.fillna(X_train1.max(),inplace=True)
X_test1.fillna(X_test1.max(),inplace=True)

lm1=linear_model.LinearRegression()
model1=lm1.fit(X_train1,Y_train1)
pred1=lm1.predict(X_test1)
pred1=pred1.round()
pred1=(np.maximum(pred1,0.))

print(model1.coef_)
print(model1.intercept_)

np.savetxt("C:/Users/user/Downloads/weibo_predict_result51.csv",pred1,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result1=pd.read_csv("C:/Users/user/Downloads/weibo_predict_result51.csv")

print(mean_squared_error(Y_test1,result1))

train_real_pred=Y_test1
train_real_pred['fp']=result1['forward_count']
train_real_pred['cp']=result1['comment_count']
train_real_pred['lp']=result1['like_count']
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model10: (Factors: Media, Length, Emoji, Median,Polarity) with OLS

### OLS is a type of linear least sqaures methods for estimating parameters in a linear regression model

X_train1=train[["content_media_count","content_length","forward_median","comment_median","like_median","polarity"]]
Y_train1=train[["forward_count","comment_count","like_count"]]
X_test1=predict[["content_media_count","content_length","forward_median","comment_median","like_median","polarity"]]
Y_test1=predict[["forward_count","comment_count","like_count"]]

pd.options.mode.use_inf_as_na = True
X_train1.fillna(X_train1.max(),inplace=True)
X_test1.fillna(X_test1.max(),inplace=True)

model1=sm.OLS(Y_train1,X_train1).fit()
pred1=model1.predict(X_test1)
pred1=pred1.round()
pred1=(np.maximum(pred1,0.))

np.savetxt("C:/Users/user/Downloads/weibo_predict_result52.csv",pred1,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result1=pd.read_csv("C:/Users/user/Downloads/weibo_predict_result52.csv")

train_real_pred=Y_test1
train_real_pred['fp']=result1['forward_count']
train_real_pred['cp']=result1['comment_count']
train_real_pred['lp']=result1['like_count']
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model11: (Factors: Media, Length, Emoji, Median,Polarity) with Ridge regression

### Ridge regression is used to prevent multicollinearity among variables by shrinking the parameters

X_train1=train[["content_media_count","content_length","forward_median","comment_median","like_median","polarity"]]
Y_train1=train[["forward_count","comment_count","like_count"]]
X_test1=predict[["content_media_count","content_length","forward_median","comment_median","like_median","polarity"]]
Y_test1=predict[["forward_count","comment_count","like_count"]]

pd.options.mode.use_inf_as_na = True
X_train1.fillna(X_train1.max(),inplace=True)
X_test1.fillna(X_test1.max(),inplace=True)

lm1=linear_model.Ridge(alpha=3)
model1=lm1.fit(X_train1,Y_train1)
pred1=lm1.predict(X_test1)
pred1=pred1.round()
pred1=(np.maximum(pred1,0.))

print(model1.coef_)
print(model1.intercept_)

np.savetxt("C:/Users/user/Downloads/weibo_predict_result53.csv",pred1,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result1=pd.read_csv("C:/Users/user/Downloads/weibo_predict_result53.csv")

train_real_pred=Y_test1
train_real_pred['fp']=result1['forward_count']
train_real_pred['cp']=result1['comment_count']
train_real_pred['lp']=result1['like_count']
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))

# Model 12: (Factors: Media, Length, Emoji, Median,Polarity) with Lasso regression

### Lasso regression does automatic feature selection that means if some features are correlated then lasso will pick only one feature

X_train1=train[["content_media_count","content_length","forward_median","comment_median","like_median","polarity"]]
Y_train1=train[["forward_count","comment_count","like_count"]]
X_test1=predict[["content_media_count","content_length","forward_median","comment_median","like_median","polarity"]]
Y_test1=predict[["forward_count","comment_count","like_count"]]

pd.options.mode.use_inf_as_na = True
X_train1.fillna(X_train1.max(),inplace=True)
X_test1.fillna(X_test1.max(),inplace=True)

lm1=Lasso(alpha=0.01)
model1=lm1.fit(X_train1,Y_train1)
pred1=lm1.predict(X_test1)
pred1=pred1.round()
pred1=(np.maximum(pred1,0.))

print(model1.coef_)
print(model1.intercept_)

np.savetxt("C:/Users/user/Downloads/weibo_predict_result54.csv",pred1,delimiter=',',header="forward_count,comment_count,like_count",comments="")
result1=pd.read_csv("C:/Users/user/Downloads/weibo_predict_result54.csv")

train_real_pred=Y_test1
train_real_pred['fp']=result1['forward_count']
train_real_pred['cp']=result1['comment_count']
train_real_pred['lp']=result1['like_count']
print("Score:{0:.2f}%".format(precision(train_real_pred.values)*100))
