==> Sina Webo Interaction Prediction/18. Gradient Boost.ipynb <==
from sklearn.ensemble import GradientBoostingRegressor

import pandas as pd
import numpy as np
from sklearn import linear_model
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from matplotlib import pyplot as plt
import statsmodels.api as sm

import import_ipynb
from evaluation import precision
from runTime import runTime

df1=pd.read_csv("E:\\5th-Sem\\DMA Project\\Project\\weibo_train1_cpts.csv")
df2=pd.read_csv("E:\\5th-Sem\\DMA Project\\Project\\weibo_train2_cpts.csv")
frames=[df1,df2]
train_all=pd.concat(frames)

train_all.columns

train=train_all[0:983694]
cv=train_all[983695:1229618]

X_train6=train[["content_media_count","forward_median","comment_median","like_median"]]
X_test6= cv[["content_media_count","forward_median","comment_median","like_median"]]
Y_train6=train[["forward_count"]]
Y_test6=cv[["forward_count"]]


gbrt=GradientBoostingRegressor(n_estimators=100) 
gbrt.fit(X_train6, Y_train6) 
y_pred=gbrt.predict(X_test6)
y_pred=y_pred.round()
y_pred=(np.maximum(y_pred,0.))

np.savetxt("E://DMA_PRE//weibo_predict_resultbow6.csv",y_pred,delimiter=',',header="forward_count",comments="")
result6=pd.read_csv("E://DMA_PRE//weibo_predict_resultbow6.csv")
result6=result6.abs()
result6=result6.astype(int)


X_train7=train[["content_media_count","forward_median","comment_median","like_median"]]
X_test7= cv[["content_media_count","forward_median","comment_median","like_median"]]
Y_train7=train[["like_count"]]
Y_test7=cv[["like_count"]]

gbrt=GradientBoostingRegressor(n_estimators=100) 
gbrt.fit(X_train7, Y_train7) 
y_pred1=gbrt.predict(X_test7)
y_pred1=y_pred1.round()
y_pred1=(np.maximum(y_pred1,0.))

np.savetxt("E://DMA_PRE//weibo_predict_resultbow7.csv",y_pred1,delimiter=',',header="like_count",comments="")
result7=pd.read_csv("E://DMA_PRE//weibo_predict_resultbow7.csv")
result7=result7.abs()
result7=result7.astype(int)

X_train8=train[["content_media_count","forward_median","comment_median","like_median"]]
X_test8= cv[["content_media_count","forward_median","comment_median","like_median"]]
Y_train8=train[["comment_count"]]
Y_test8=cv[["comment_count"]]

gbrt=GradientBoostingRegressor(n_estimators=100) 
gbrt.fit(X_train8, Y_train8) 
y_pred2=gbrt.predict(X_test8)
y_pred2=y_pred2.round()
y_pred2=(np.maximum(y_pred2,0.))

np.savetxt("E://DMA_PRE//weibo_predict_resultbow8.csv",y_pred2,delimiter=',',header="comment_count",comments="")
result8=pd.read_csv("E://DMA_PRE//weibo_predict_resultbow8.csv")
result8=result8.abs()
result8=result8.astype(int)


train_real_pred = pd.concat([Y_test6,Y_test8,Y_test7],axis=1)
train_real_pred['fp']=result6['forward_count'].values
train_real_pred['cp']=result8['comment_count'].values
train_real_pred['lp']=result7['like_count'].values
train_real_pred=train_real_pred.round()
print ("Score on the training set:{0:.2f}%".format(precision(train_real_pred.values)*100))
